style: diaryqueryservice로 합치기 (#48)
